Adds some cluster love, next up i2c

+++ adc.c
@@ -53,7 +53,7 @@ void datn_adc_init(){
 }
 
 /* 32 channels at maximum */
-void datn_adc_sample(uint16_t *channels, uint32_t channels_to_sample) {
+void datn_adc_sample(uint32_t *channels, uint32_t channels_to_sample) {
     for (uint32_t i = 0; i < channels_to_sample; i++)
     {
         select_channel(i);
