+++ adc.h
@@ -6,6 +6,6 @@
 void datn_adc_init();
 
 /* 32 channels at maximum */
-void datn_adc_sample(uint16_t *channels, uint32_t channels_to_sample);
+void datn_adc_sample(uint32_t *channels, uint32_t channels_to_sample);
 
 #endif // ADC_H_
